Correções e inicio dos testes

+++ Core/Repository/IContatoRepository.cs
@@ -8,9 +8,9 @@ namespace Core.Repository
     public interface IContatoRepository: IDisposable
     {
         IEnumerable<Contato> ToList();
-        Contato GetByID(int Id);
+        Contato GetByID(long Id);
         void Insert(Contato student);
-        void Delete(int id);
+        void Delete(long id);
         void Update(Contato contato);
         void Save();
     }
